requirements.txt
# Core ML libraries
lightgbm>=4.0.0
scikit-learn>=1.3.0
shap>=0.42.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# API and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# UI
streamlit>=1.28.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0

# LLM integration
openai>=1.0.0
google-generativeai>=0.3.0

# Testing
pytest>=7.0.0

# Future dependencies (commented out for now)
# econml>=0.14.0
# causalml>=0.4.0
# litellm>=1.0.0
# pgvector>=0.2.0
